+++ validation/opto_val.py
@@ -93,14 +93,14 @@ def get_Irr0_thres(
                 # use 240*(thresh-E_L) factor from tutorial
                 Iopto_per_mW_per_mm2=Iopto_gain_from_factor(gain_factor)
             ),
-            light_model_params=default_blue,
+            light_model_params=fiber_params_blue,
             location=(0, 0, 0) * mm,
         )
     else:
         opto = OptogeneticIntervention(
             name="opto",
             opsin_model=FourStateModel(ChR2_four_state),
-            light_model_params=default_blue,
+            light_model_params=fiber_params_blue,
             location=(0, 0, 0) * mm,
         )
     sim.inject_stimulator(opto, ng)
